bpf: introduce new bpf map type BPF_MAP_TYPE_WILDCARD
Add a new map, BPF_MAP_TYPE_WILDCARD, which provides means to implement generic
online packet classification. Here "online" stands for "fast lookups and fast
updates", and "generic" means that a user can create maps with custom lookup
schemes, i.e., with different numbers of fields and different interpretation
of individual fields (prefix bitmasks, ranges, matches, and wildcard matches).

The original intent to add this map was to implement Network Policies for
cilium which would support port ranges and to address some lower priority
goals, namely, to improve functionality of cilium XDP Prefilter and cilium
Packet Recorder. The map implementation uses a modified Tuple Merge algorithm
(see [1], [2]).

The map supports a set of operations similar to the hashmap, however, the
bpf_map_lookup_elem helper can be used for both lookup rules and match input
against existing rules. (All the other operations are performed only on rules.)
This is implemented as follows. Keys for wildcard maps are defined using the
struct wildcard_key structure:

    struct wildcard_key {
            __u32 type;
            __u32 priority;
            __u8 data[];
    };

Here the type filed is used to distinguish between rules and matches. To modify
or lookup a rule, the type field should be set to BPF_WILDCARD_KEY_RULE. To
match input the type field should be set to BPF_WILDCARD_KEY_MATCH. The
priority field may be used to specify the order of precedence in which rules
should be matched (lower value => higher precedence).

The data[] field is a map-specific (packed) structure which should fit both
rule descriptions and input data (the latter is always of the same size or
smaller than the rule size). Every map instance can be configured to have
different number of rules. Rules can be of several types: prefix, range, match,
and wildcard match. There can be up to 9 different rules per map. Examples of
configuration (every example represents a separate map):

    * [32-bit match, 16-bit range] (say, ID + destination port)
    * [128-bit prefix, 16-bit range, 8-bit wildcard match] (say, IPv6 CIDR + destination port + protocol)
    * [32-bit prefix, 32-bit prefix, 16-bit range, 16-bit range] (say, IPv4 4-tuple)

Given a particular structure, the data[] field should contain enough room to
fit the description. This is how different rules (of size SIZE) are stored in
the data[] field:

    * prefix: value, u32 prefix
    * range: min_value, max_value
    * match: value
    * wildcard match: value

If there is more than one rule, then they are packed together. To do a match,
the data[] field should only contain (packed) values.

To simplify life of a user, a set of macros is added to the uapi/linux/bpf.h
file which can be used to define proper structures. For example, for the
ID+dport example above, one could use the following macro:

    BPF_WILDCARD_DESC_2(
            policy,
            BPF_WILDCARD_RULE_PREFIX, __u32, id,
            BPF_WILDCARD_RULE_RANGE, __u16, dport
    );

This macro defines the following structure (ignore the desc[0] field for
a second):

    struct policy_key {
            struct policy_desc desc[0];
            __u32 type;
            __u32 priority;
            union {
                    struct {
                            __u32 id;
                            __u16 dport_min;
                            __u16 dport_max;
                    } __attribute__((packed)) rule;
                    struct {
                            __u32 id;
                            __u16 dport;
                    } __attribute__((packed));
            };
    } __attribute__((packed));

So to configure a rule one can do, e.g.,

    struct policy_key key = {
            .type = BPF_WILDCARD_KEY_RULE,
            .rule.id = 0xbeef,
            .rule.dport_min = 1,
            .rule.dport_max = 1024,
    };
    bpf_map_update_elem(map, &key, &val);

and to match the input, one can do

    struct policy_key key = {
            .type = BPF_WILDCARD_KEY_MATCH,
            .id = 0xbeef,
            .dport = 557,
    };
    bpf_map_lookup_elem(map, &key);

Now, the magic field desc[0] is not a typo, but a requirement. The BTF
description of the wildcard map must contain this field, and it should be of
the following format:

    struct {
            __uint(n_rules, N);
            __uint(rule_1, TYPE_1 << 8 | SIZE_1);
            __uint(rule_2, TYPE_2 << 8 | SIZE_2);
            ...
            __uint(rule_N, TYPE_N << 8 | SIZE_N);
    };

In our case the exact format is

    struct policy_desc {
            __uint(n_rules, 2);
            __uint(id, BPF_WILDCARD_RULE_MATCH << 8 | sizeof(__u32));
            __uint(dport, BPF_WILDCARD_RULE_RANGE << 8 | sizeof(__u16));
    };

The structure 'struct policy_desc' is defined by the same BPF_WILDCARD_DESC_2
macro, so the full map definition would be just

    BPF_WILDCARD_DESC_2(
            policy,
            BPF_WILDCARD_RULE_PREFIX, __u32, id,
            BPF_WILDCARD_RULE_RANGE, __u16, dport
    );
    struct {
            __uint(type, BPF_MAP_TYPE_WILDCARD);
            __type(key, struct policy_key);
            __type(value, __u64);
            __uint(max_entries, MAX_ENTRIES);
            __uint(map_flags, BPF_F_NO_PREALLOC);
    } NAME SEC(".maps")

A wildcard map can't be created without passing a btf_fd and a proper key and
value BTF ids.

Links:
  [1] "TupleMerge: Fast Software Packet Processing for Online Packet Classification", https://nonsns.github.io/paper/rossi19ton.pdf
  [2] "A BPF map for online packet classification", https://lpc.events/event/16/contributions/1356/

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>