selftest/bpf/benchs: make argp enums global
Currently every benchmark which has its own command line options define a enum
with option ids, e.g.,

    enum {
           ARG_SMTH = 9000,
           ARG_SMTH_ELSE = 9001,
    };

These values are later used to distinguish between command line options:

    static const struct argp_option opts[] = {
            { "smth", ARG_SMTH, "SMTH", 0,
                    "Set number of smth to configure smth"},
            { "smth_else", ARG_SMTH_ELSE, "SMTH_ELSE", 0,
                    "Set number of smth_else to configure smth else"},
            ...

If another benchmark wants to use an argument with the same name, then it
fails, because this is not determined which of definitions will be applied when
parsing arguments.

Move arguments id definitions to bench.h such that they are visible to every
benchmark (and such that there's no need to grep if this number is defined
already or not). Arguments with same names should have the same semantics
(taking parameter or not), but can have different description and different
parsers.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>