lib: add a new version of xxh which works faster for small keys
Add a new modern XXH3 function by Yann Collet [1].  The kernel variant only
suports inputs of lengths <= 240.  The reason for that is that the original
XXH3 was designed and optimized to get most benefits from using vectorized
instructions, and the scalar version for inputs of lengths > 240 looses to
the older xxh* functions significantly.  For the ease of use, combine the
two functions into one: xxh_combined.

TBD: document tests and benchmarks

  [1] the original xxh* code: https://github.com/Cyan4973/xxHash

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>