[draft] bpf: replace hash in hashtab by xxh_combined
TBD: Based on the benchmarks it is supposed to improve the lookup speeds (as
the jhash2 looses when key > 16). Instead, we're getting above only for keys >
32 or so. Why? Compiling xxh as O3 also doesn't help. Besides that flame graph
says that we're spending more than half time in lookup_nulls_elem_raw(), which
means that for keys <40-60 we don't care about the hash too much.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>