bpf: bind extra maps to programs on load
Add new bind_fd_array and bind_fd_array_cnt attributes to the
BPF_PROG_LOAD command to pass an array of file descriptors to be
bound to the program being loaded. This functionality is similar to
the functionality provided by the BPF_PROG_BIND_MAP syscall besides
that this is happening on program load and thus the maps may be used
by the verifer.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>