bpf: add fd_array_cnt attribute for prog_load
The fd_array attribute of the BPF_PROG_LOAD syscall may contain a set
of file descriptors: maps or btfs. This field was introduced as a
sparse array. Introduce a new attribute, fd_array_cnt, which, if
present, indicates that the fd_array is a continuous array of the
corresponding length.

If fd_array_cnt is non-zero, then every map in the
fd_array will be bound to the program, as if it was used by the
program. This functionality is similar to the BPF_PROG_BIND_MAP
syscall, but such maps can be used by the verifier during the program
load.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>