TBD: add support for indirect jumps in BPF
TBD: correct the commit message below; it was written for the previous version of the instruciton

.....

Add support for a new instruction

    BPF_JMP|BPF_X|BPF_JA, SRC=Rx, DST=0, off=0, imm=fd(M)

which does an indirect jump to a location stored in M[Rx]. Here M must be a
corresponding instruction set map (see TBD). This instruction is replaced by
verifier with

    Rt = map-address(M)
    Rt += offset-to-map-elements
    Rx *= elem_size(M)
    Rx += Rt

so Rx now points to a corresponding jitted_ip field in M.

In future a different variant of this instruction will be added (DST=1):

    BPF_JMP|BPF_X|BPF_JA, SRC=Rx, DST=1, off=0, imm=fd(M)

which will issue an indirect jump to a location pointed by Rx: jmp *(rX). This
requires way more verification, such as read-only maps, etc., and TBD as a
consequent patch.

Signed-off-by: Anton Protopopov <a.s.protopopov@gmail.com>