bpf: Add support for static branch instructions
Add support for a new version of JA instruction, a static branch JA.
Such instructions may either jump to the specified offset or be nops.
To distinguish them from normal JA the new BPF_STATIC_BRANCH_JA flag
should be set in the SRC register.

By default, on a program load such instructions are jitted as jumps.
However, if the BPF_STATIC_BRANCH_NOP flag is set in the SRC register,
then the instruction is jitted to a nop.

In order to generate BPF_STATIC_BRANCH_JA instructions using llvm two
new instructions will be added:

    asm volatile goto ("nop_or_gotol %l[label]" :::: label);

will generate the BPF_STATIC_BRANCH_JA|BPF_STATIC_BRANCH_NOP
instruction and

    asm volatile goto ("gotol_or_nop %l[label]" :::: label);

will generate a BPF_STATIC_BRANCH_JA instruction. The reason for
adding two instructions is that both are required to implement
static keys functionality for BPF, namely, to be able to implement
the "likely" and "unlikely" flavours of branches.

Signed-off-by: Anton Protopopov <a.s.protopopov@gmail.com>